+++ OneSignalExample/Assets/OneSignal/CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - iOS crash when calling OneSignal.User.PushSubscription.Id and OneSignal.User.PushSubscription.Token when they are null.
 
+## [5.1.0]
+### Fixed
+- Add getters for the user state that includes onesignalId and externalId
+- Add observer for the user state
+### Changed
+- Updated included Android SDK to [5.1.4](https://github.com/OneSignal/OneSignal-Android-SDK/releases/tag/5.1.4)
+
 ## [5.0.6]
 ### Fixed
 - Duplicate symbol errors when building with other iOS plugins